Se agrega al view la funcion para mostrar la lista de valores ausentes. Se oculta la carpeta de la base de datos.

+++ csv_uploader/models.py
@@ -6,18 +6,18 @@ class CSVData(models.Model):
     stock_code = models.CharField(max_length=10)
     description = models.CharField(max_length=255)
     quantity = models.IntegerField()
-    unit_price = models.FloatField()
+    unit_price = models.DecimalField(max_digits=10, decimal_places=2)
     customer_id = models.CharField(max_length=10)
     region = models.CharField(max_length=100)
     invoice_day = models.IntegerField()
     invoice_month = models.IntegerField()
     invoice_year = models.IntegerField()
     invoice_hour = models.IntegerField()
-    total_sales = models.FloatField()
-    sales_per_customer = models.FloatField()
-    invoices_per_customer = models.FloatField()
+    total_sales = models.DecimalField(max_digits=10, decimal_places=2)
+    sales_per_customer = models.DecimalField(max_digits=10, decimal_places=2)
+    invoices_per_customer = models.DecimalField(max_digits=10, decimal_places=2)
     avg_sales_per_customer = models.FloatField()
-    neg_invoices_per_customer = models.FloatField()
+    neg_invoices_per_customer = models.IntegerField()
 
     def __str__(self):
         return self.invoice_no
